Add a playRound function
I made the main JavaScript functionality of my rock paper scissors game.
Once I was able to collect, the computer's randomized choice. I'm able
to compare it to the user's choice for a proper round of
rock-paper-scissors. I simply added a new function called playRound that
allows me to play a round of rock-paper-scissors.