Add game function and update playRound
I created a game function that allows the player to play against the
computer a best of 5 rounds game while disregarding ties as rounds. Once
the computer and player play five rounds, the console prints out the
final winner of the game. I had to update playRound to accomodate for
the game function since I wanted to deal with the error handling of user
input and forced replayability for tied rounds. Thus, I moved the
case filtering to the game function and used tie's return value to stop
the round counter from incrementing. I might rewrite this code again for
more modularity and cleanliness.